Use `union` as a no-op grouping command
We need a no-op modifier as otherwise attempting to apply a debug
modifier to a bracketed group of commands fails.  Use `union` as the
no-op modifier, rather than `rotate([0,0,0])`, on the grounds that the
former is always implied by any grouping operation where something else
isn't specified, and therefore feels (at least to me) more correct for
the purpose.