Support non-Docker builds
Default to building with a local OpenSCAD installation, or use Docker if
USE_DOCKER=Yes is given to make.

The point of Docker is to allow this to be compiled by folk who have
Docker installed but not OpenSCAD.  But it's surely going to be more
common for anyone wanting to compile this to have OpenSCAD but not
Docker, so support OpenSCAD as the default, and make using Docker a
compile-time option.